use determined hosted mnist dataset
