+++ examples/computer_vision/_mnist_pl/mnist.py
@@ -1,16 +1,7 @@
 import torch
 import pytorch_lightning as pl
-from torch.utils.data import DataLoader
 from torch.nn import functional as F
-from torchvision import transforms
-from torchvision.datasets import MNIST
-from pathlib import Path
-from typing import Optional
-import os
-import urllib.parse
-import requests
-import logging
-import shutil
+from data import MNISTDataModule
 
 
 class LightningMNISTClassifier(pl.LightningModule):
@@ -23,6 +14,10 @@ class LightningMNISTClassifier(pl.LightningModule):
         self.layer_1 = torch.nn.Linear(28 * 28, 128)
         self.layer_2 = torch.nn.Linear(128, 256)
         self.layer_3 = torch.nn.Linear(256, 10)
+        # self.dims is returned when you call dm.size()
+        # Setting default dims here because we know them.
+        # Could optionally be assigned dynamically in dm.setup()
+        self.dims = (1, 28, 28)
 
 
     def forward(self, x):
@@ -72,46 +67,6 @@ class LightningMNISTClassifier(pl.LightningModule):
                                      lr=self.hparams.lr)
         return optimizer
 
-
-class MNISTDataModule(pl.LightningDataModule):
-    def __init__(self, data_url: str, data_dir: str = '/tmp/det'):
-        super().__init__()
-        self.data_dir = data_dir
-        self.data_url = data_url
-
-    def prepare_data(self):
-        # prepare transforms standard to MNIST
-        url_path = urllib.parse.urlparse(self.data_url).path
-        basename = url_path.rsplit("/", 1)[1]
-
-        download_directory = os.path.join(self.data_dir, "MNIST")
-        os.makedirs(download_directory, exist_ok=True)
-        filepath = os.path.join(download_directory, basename)
-        if not os.path.exists(filepath):
-            logging.info("Downloading {} to {}".format(self.data_url, filepath))
-
-            r = requests.get(self.data_url, stream=True)
-            with open(filepath, "wb") as f:
-                for chunk in r.iter_content(chunk_size=8192):
-                    if chunk:
-                        f.write(chunk)
-
-        shutil.unpack_archive(filepath, download_directory)
-
-        self.data_dir = os.path.dirname(download_directory)
-
-    def setup(self, stage: Optional[str] = None):
-        transform = transforms.Compose([transforms.ToTensor(),
-                                        transforms.Normalize((0.1307,), (0.3081,))])
-
-        # prepare transforms standard to MNIST
-        self.mnist_train = MNIST(str(Path(self.data_dir)), train=True, transform=transform)
-        self.mnist_val = MNIST(str(Path(self.data_dir)), train=False, transform=transform)
-
-    def train_dataloader(self):
-        return DataLoader(self.mnist_train, batch_size=64, num_workers=12)
-    def val_dataloader(self):
-        return DataLoader(self.mnist_val, batch_size=64)
 
 if __name__ == '__main__':
     model = LightningMNISTClassifier(lr=1e-3)
