+++ examples/gan/_gan_mnist_pl/adaptive.yaml
@@ -1,4 +1,6 @@
 description: gan_mnist_pytorch_lightning_adaptive
+data:
+  url: "https://s3-us-west-2.amazonaws.com/determined-ai-test-data/pytorch_mnist.tar.gz"
 records_per_epoch: 55000
 hyperparameters:
   global_batch_size: 64
