+++ examples/gan/_gan_mnist_pl/gan.py
@@ -18,8 +18,6 @@ Main takeaways:
 Lightning is easy to install. Simply `pip install pytorch-lightning`
 """
 
-import os
-from argparse import ArgumentParser
 from collections import OrderedDict
 
 import numpy as np
@@ -27,56 +25,10 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 import torchvision
-import torchvision.transforms as transforms
-from torch.utils.data import DataLoader, random_split
-from torchvision.datasets import MNIST
 
 import pytorch_lightning as pl
+from data import MNISTDataModule
 
-
-class MNISTDataModule(pl.LightningDataModule):
-
-    def __init__(self, data_dir: str = '/tmp/gan-mnist', batch_size: int = 64, num_workers: int = 8):
-        super().__init__()
-        self.data_dir = data_dir
-        self.batch_size = batch_size
-        self.num_workers = num_workers
-
-        self.transform = transforms.Compose([
-            transforms.ToTensor(),
-            transforms.Normalize((0.1307,), (0.3081,))
-        ])
-
-        # self.dims is returned when you call dm.size()
-        # Setting default dims here because we know them.
-        # Could optionally be assigned dynamically in dm.setup()
-        self.dims = (1, 28, 28)
-        self.num_classes = 10
-
-    def prepare_data(self):
-        # download
-        MNIST(self.data_dir, train=True, download=True)
-        MNIST(self.data_dir, train=False, download=True)
-
-    def setup(self, stage=None):
-
-        # Assign train/val datasets for use in dataloaders
-        if stage == 'fit' or stage is None:
-            mnist_full = MNIST(self.data_dir, train=True, transform=self.transform)
-            self.mnist_train, self.mnist_val = random_split(mnist_full, [55000, 5000])
-
-        # Assign test dataset for use in dataloader(s)
-        if stage == 'test' or stage is None:
-            self.mnist_test = MNIST(self.data_dir, train=False, transform=self.transform)
-
-    def train_dataloader(self):
-        return DataLoader(self.mnist_train, batch_size=self.batch_size, num_workers=self.num_workers)
-
-    def val_dataloader(self):
-        return DataLoader(self.mnist_val, batch_size=self.batch_size, num_workers=self.num_workers)
-
-    def test_dataloader(self):
-        return DataLoader(self.mnist_test, batch_size=self.batch_size, num_workers=self.num_workers)
 
 
 class Generator(nn.Module):
@@ -247,7 +199,7 @@ class GAN(pl.LightningModule):
 
 
 if __name__ == '__main__':
-    dm = MNISTDataModule()
+    dm = MNISTDataModule('https://s3-us-west-2.amazonaws.com/determined-ai-test-data/pytorch_mnist.tar.gz')
     model = GAN(*dm.size())
     trainer = pl.Trainer(gpus=1, max_epochs=5, progress_bar_refresh_rate=20, default_root_dir='/tmp/lightning')
     trainer.fit(model, dm)
