+++ examples/gan/_gan_mnist_pl/model_def.py
@@ -10,8 +10,8 @@ import gan
 
 class GANTrial(PLAdapter):
     def __init__(self, context: PyTorchTrialContext) -> None:
-        data_dir = f"/tmp/data-rank{context.distributed.get_rank()}"
-        self.dm = gan.MNISTDataModule(data_dir,
+        data_dir = f'/tmp/data-rank{context.distributed.get_rank()}'
+        self.dm = gan.MNISTDataModule(context.get_data_config()['url'], data_dir,
                                       batch_size=context.get_global_batch_size())
         channels, width, height = self.dm.size()
         lm = gan.GAN(channels, width, height,
